NETStandard releated improvelents + another websocket fix
