+++ Titanium.Web.Proxy/Extensions/TcpExtensions.cs
@@ -34,6 +34,7 @@ namespace Titanium.Web.Proxy.Extensions
             }
         }
 
+#if NET45
         /// <summary>
         /// Gets the local port from a native TCP row object.
         /// </summary>
@@ -53,5 +54,6 @@ namespace Titanium.Web.Proxy.Extensions
         {
             return (tcpRow.remotePort1 << 8) + tcpRow.remotePort2 + (tcpRow.remotePort3 << 24) + (tcpRow.remotePort4 << 16);
         }
+#endif
     }
 }
