+++ Titanium.Web.Proxy/Helpers/SystemProxy.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Linq;
-using System.Runtime.InteropServices;
 using Microsoft.Win32;
 
 // Helper classes for setting system proxy settings
@@ -28,24 +27,6 @@ namespace Titanium.Web.Proxy.Helpers
         /// Both HTTP and HTTPS
         /// </summary>
         AllHttp = Http | Https,
-    }
-
-    internal partial class NativeMethods
-    {
-        [DllImport("wininet.dll")]
-        internal static extern bool InternetSetOption(IntPtr hInternet, int dwOption, IntPtr lpBuffer, int dwBufferLength);
-
-        [DllImport("kernel32.dll")]
-        internal static extern IntPtr GetConsoleWindow();
-
-        // Keeps it from getting garbage collected
-        internal static ConsoleEventDelegate Handler;
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        internal static extern bool SetConsoleCtrlHandler(ConsoleEventDelegate callback, bool add);
-
-        // Pinvoke
-        internal delegate bool ConsoleEventDelegate(int eventType);
     }
 
     internal class HttpSystemProxyValue
