+++ Titanium.Web.Proxy/Helpers/Tcp.cs
@@ -1,11 +1,9 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Net.NetworkInformation;
 using System.Runtime.InteropServices;
 using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 using Titanium.Web.Proxy.Extensions;
 using Titanium.Web.Proxy.Models;
@@ -20,63 +18,9 @@ namespace Titanium.Web.Proxy.Helpers
         Ipv6 = 2,
     }
 
-    internal partial class NativeMethods
-    {
-        internal const int AfInet = 2;
-        internal const int AfInet6 = 23;
-
-        internal enum TcpTableType
-        {
-            BasicListener,
-            BasicConnections,
-            BasicAll,
-            OwnerPidListener,
-            OwnerPidConnections,
-            OwnerPidAll,
-            OwnerModuleListener,
-            OwnerModuleConnections,
-            OwnerModuleAll,
-        }
-
-        /// <summary>
-        /// <see href="http://msdn2.microsoft.com/en-us/library/aa366921.aspx"/>
-        /// </summary>
-        [StructLayout(LayoutKind.Sequential)]
-        internal struct TcpTable
-        {
-            public uint length;
-            public TcpRow row;
-        }
-
-        /// <summary>
-        /// <see href="http://msdn2.microsoft.com/en-us/library/aa366913.aspx"/>
-        /// </summary>
-        [StructLayout(LayoutKind.Sequential)]
-        internal struct TcpRow
-        {
-            public TcpState state;
-            public uint localAddr;
-            public byte localPort1;
-            public byte localPort2;
-            public byte localPort3;
-            public byte localPort4;
-            public uint remoteAddr;
-            public byte remotePort1;
-            public byte remotePort2;
-            public byte remotePort3;
-            public byte remotePort4;
-            public int owningPid;
-        }
-
-        /// <summary>
-        /// <see href="http://msdn2.microsoft.com/en-us/library/aa365928.aspx"/>
-        /// </summary>
-        [DllImport("iphlpapi.dll", SetLastError = true)]
-        internal static extern uint GetExtendedTcpTable(IntPtr tcpTable, ref int size, bool sort, int ipVersion, int tableClass, int reserved);
-    }
-
     internal class TcpHelper
     {
+#if NET45
         /// <summary>
         /// Gets the extended TCP table.
         /// </summary>
@@ -165,6 +109,7 @@ namespace Titanium.Web.Proxy.Helpers
 
             return null;
         }
+#endif
 
         /// <summary>
         /// relays the input clientStream to the server at the specified host name and port with the given httpCmd and headers as prefix
@@ -202,7 +147,7 @@ namespace Titanium.Web.Proxy.Helpers
                     writer.Flush();
 
                     var data = ms.ToArray();
-                    await ms.WriteAsync(data, 0, data.Length);
+                    await clientStream.WriteAsync(data, 0, data.Length);
                     onDataSend?.Invoke(data, 0, data.Length);
                 }
             }
